requirements.txt
# AWS / Bedrock
boto3>=1.34.0
botocore>=1.34.0

# Web framework
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
websockets>=12.0
pydantic>=2.0.0

# Voice / Audio
openai-whisper>=20231117
sounddevice>=0.4.6
numpy>=1.24.0
scipy>=1.11.0

# Config
pyyaml>=6.0.1

# Vector memory
chromadb>=0.4.22
sentence-transformers>=2.6.0

# Utilities
rich>=13.7.0
httpx>=0.27.0
